Refine

+++ fanto/fanto/viewcontrollers/profile/learning/FTCongratsViewController.m
@@ -67,7 +67,7 @@
 
 #pragma mark - FTActionSheetDelegate methods
 - (void)actionSheetDidSelectAtIndex:(NSInteger)index {
-  DLog(@"%d", index);
+  DLog(@"%ld", (long)index);
 }
 
 @end
